+++ Tests/TestCase.swift
@@ -29,9 +29,7 @@ extension CoreDataStackProvider {
 
     func createStack() -> CoreDataStack.StackResult {
         var result: CoreDataStack.StackResult!
-        createStack(onQueue: nil) { r in
-            result = r
-        }
+        self.createStack(onQueue: nil) { result = $0 }
         return result
     }
 }
@@ -52,13 +50,13 @@ class TestCase: XCTestCase {
     override func setUp() {
         super.setUp()
 
-        let factory = CoreDataStackProvider(model: inMemoryModel)
+        let factory = CoreDataStackProvider(model: self.inMemoryModel)
         let result = factory.createStack()
-        inMemoryStack = try! result.get()
+        self.inMemoryStack = try! result.get()
     }
 
     override func tearDown() {
-        inMemoryStack = nil
+        self.inMemoryStack = nil
         super.tearDown()
     }
 
@@ -68,10 +66,10 @@ class TestCase: XCTestCase {
     func generateDataInContext(_ context: NSManagedObjectContext,
                                companiesCount: Int = Int.random(in: 0...10),
                                employeesCount: Int = Int.random(in: 0...1_000)) -> [Company] {
-        let companies = generateCompaniesInContext(context, count: companiesCount)
+        let companies = self.generateCompaniesInContext(context, count: companiesCount)
 
-        companies.forEach { c in
-            generateEmployeesInContext(context, count: employeesCount, company: c)
+        companies.forEach { company in
+            self.generateEmployeesInContext(context, count: employeesCount, company: company)
         }
 
         return companies
@@ -82,8 +80,8 @@ class TestCase: XCTestCase {
         var companies = [Company]()
 
         for _ in 0..<count {
-            let c = Company.newCompany(context)
-            companies.append(c)
+            let company = Company.newCompany(context)
+            companies.append(company)
         }
 
         return companies
@@ -94,8 +92,8 @@ class TestCase: XCTestCase {
         var employees = [Employee]()
 
         for _ in 0..<count {
-            let c = Employee.newEmployee(context, company: company)
-            employees.append(c)
+            let company = Employee.newEmployee(context, company: company)
+            employees.append(company)
         }
 
         return employees