+++ scripts/build_docs.zsh
@@ -5,12 +5,10 @@
 #
 #  Copyright © 2020-present Jesse Squires
 #
-#  Docs by jazzy
-#  https://github.com/realm/jazzy/releases/latest
-#  ------------------------------
+#  Jazzy: https://github.com/realm/jazzy/releases/latest
 #  Generates documentation using jazzy and checks for installation.
 
-VERSION="0.13.6"
+VERSION="0.13.7"
 
 FOUND=$(jazzy --version)
 LINK="https://github.com/realm/jazzy"