+++ scripts/lint.zsh
@@ -5,19 +5,54 @@
 #
 #  Copyright © 2020-present Jesse Squires
 #
-#  SwiftLint
-#  https://github.com/realm/SwiftLint/releases/latest
-#  ------------------------------
-#  Runs SwiftLint and checks for installation.
+#  SwiftLint: https://github.com/realm/SwiftLint/releases/latest
+#
+#  Runs SwiftLint and checks for installation of correct version.
 
-VERSION="0.41.0"
+PROJECT="JSQCoreDataKit.xcodeproj"
+SCHEME="JSQCoreDataKit"
+
+VERSION="0.43.1"
 
 FOUND=$(swiftlint version)
 LINK="https://github.com/realm/SwiftLint"
 INSTALL="brew install swiftlint"
 
+CONFIG="./.swiftlint.yml"
+
 if which swiftlint >/dev/null; then
-    swiftlint lint --config ./.swiftlint.yml
+    echo "Running swiftlint..."
+    echo ""
+
+    # no arguments, just lint without fixing
+    if [[ $# -eq 0 ]]; then
+        swiftlint --config $CONFIG
+        echo ""
+    fi
+
+    for argval in "$@"
+    do
+        # run --fix
+        if [[ "$argval" == "fix" ]]; then
+            echo "Auto-correcting lint errors..."
+            echo ""
+            swiftlint --fix --config $CONFIG && swiftlint --config $CONFIG
+            echo ""
+        # run analyze
+        elif [[ "$argval" == "analyze" ]]; then
+            LOG="xcodebuild.log"
+            echo "Running anaylze..."
+            echo ""
+            xcodebuild -scheme $SCHEME -project $PROJECT clean build-for-testing > $LOG
+            swiftlint analyze --fix --format --strict --config $CONFIG --compiler-log-path $LOG
+            rm $LOG
+            echo ""
+        else
+            echo "Error: invalid arguments."
+            echo "Usage: $0 [fix] [analyze]"
+            echo ""
+        fi
+    done
 else
     echo "
     Error: SwiftLint not installed!